Add files via upload
This project generates a Word Cloud from the IMDb movie reviews dataset to visualize the most frequently occurring words in the reviews. The dataset contains user-generated movie reviews, which are preprocessed to remove stopwords, punctuation and irrelevant text. The cleaned text is then used to create a visually appealing word cloud that highlights the most common terms, giving quick insights into audience sentiment and recurring themes in the reviews.